chore: added comments and refactor
